+++ app/assets/stylesheets/application.scss
@@ -14,5 +14,5 @@
  */
  @import "bootstrap";
  @import "./reset";
- @import "./boxes";
+ @import "./shoes";
  @import "./signup.scss";
