+++ app/assets/stylesheets/shoes.scss
@@ -1,0 +1,100 @@
+
+
+html, body {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+.container {
+  text-align:center;
+  margin: 30px auto;
+}
+
+#panel {
+  /* width: 600px;
+  height: 600px; */
+  overflow: hidden;
+  /* background-color:#000; */
+  border-radius :7px;
+  /* box-shadow: 1px 1px 3px rgba(0,0,0, .5); */
+  margin:0 auto;
+  /* cursor: move; */
+}
+
+/* #panel:hover{ */
+  /* cursor: url('img/21470.jpg'),url('img/21470.ai'); */
+  /* cursor: url("./img/aaa.png"); */
+/* } */
+
+.box {
+  width: 100px;
+  height: 100px;
+  line-height: 80px;
+  border: 1px solid silver;
+  border-radius: 10px;
+  text-align: center;
+  font-size: 26px;
+  font-weight: bold;
+  box-shadow: gray 2px 2px;
+  /* background: white; */
+  float: left;
+  /* cursor: pointer; */
+  /* cursor: url('img/21470.jpg'),url('img/21470.ai'); */
+  /* cursor: url("./img/aaa.png"); */
+  /* color: #fff; */
+  // margin: 10px 0 10px 0;
+  padding: 15px;
+  border: 9px solid #a60;
+  border-radius: 3px;
+  /* box-shadow: 2px 2px 4px #666, 2px 2px 2px #111 inset; */
+  /* text-shadow: 0px 0px 2px #000; */
+  line-height: 1.9;
+}
+
+.box:hover {
+  background-color: pink;
+  /* cursor: url('img/21470.jpg'),url('img/21470.ai'); */
+  /* cursor: url("./img/aaa.png"); */
+}
+
+#input{
+  width: 50px;
+}
+
+.btn{
+  /* width:50%; */
+}
+
+/* 裏の状態 */
+.in {
+  /* background-image: url('/img/kutsu.jpg'); */
+  background-color:#000;
+  background-size: 60px 80px;
+}
+/* 終了時 */
+.finish {
+  opacity: 0;
+  /* cursor: default; */
+}
+/* 経過秒数 */
+#result {
+  font-size: 1.2em;
+  font-weight: bold;
+}
+
+.cursor {
+  width: 20px;
+  height: 20px;
+  position: absolute;
+  top: 0;
+  left: 0; 
+  z-index: 1001;
+  cursor: none;
+  pointer-events: none;
+}
+
+img {
+  width: 100%;
+}
+
