+++ app/controllers/posts_controller.rb
@@ -1,0 +1,2 @@
+class PostsController < ApplicationController
+end
