+++ app/controllers/shoes_controller.rb
@@ -1,0 +1,2 @@
+class ShoesController < ApplicationController
+end
