+++ app/controllers/signup_controller.rb
@@ -35,14 +35,11 @@ class SignupController < ApplicationController
   private
     def admin_params
       params.require(:admin).permit(
-        :nickname,
         :email,
-        :password,
-        :password_confirmation,
-        :last_name,
-        :first_name,
-        :last_name_kana,
-        :first_name_kana
+        :encrypted_password,
+        :reset_password_token,
+        :reset_password_sent_at,
+        :remember_created_at
       )
     end
 end
