+++ app/helpers/posts_helper.rb
@@ -1,0 +1,2 @@
+module PostsHelper
+end
