+++ app/helpers/shoes_helper.rb
@@ -1,0 +1,2 @@
+module ShoesHelper
+end
