+++ app/views/shoes/index.html.erb
@@ -1,0 +1,20 @@
+
+<div class=container>
+<div class="cursor"></div>
+<div class="cursor"><img src="https://www.evoworx.co.jp/wp2/wp-content/themes/evoworx2017/common/image/dog_red.png" alt=""></div>
+
+<input type="text" id="input input-height" placeholder="上から何段目?">
+<input type="text" id="input input-width" placeholder="左から何列目?">
+<button id="go" class="btn btn-lg btn-primary">GO</button>
+<button id="on" class="btn btn-lg btn-primary">ON</button>
+<button id="search" class="btn btn-lg btn-primary">SEARCH</button>
+<button id="reset" class="btn btn-lg btn-primary">RESET</button>
+<div id="panel">
+    <!-- <div class="box empty"></div> -->
+</div>
+<div id="img_unit">
+</div>
+<div id="result"></div>
+</div>
+
+<script src="http://code.jquery.com/jquery-1.11.1.min.js"></script>
