+++ app/views/signup/step2.html.erb
@@ -1,1 +1,2 @@
-<%= @step %>
+= form_for @admin, url: signup_index_path, method: :post do |f|
+
